+++ src/add_years.js
@@ -1,4 +1,5 @@
 var parse = require('./parse')
+var getTimeSinceMidnight = require('./get_time_since_midnight')
 
 /**
  * Adds specified number of years to passed date.
@@ -8,7 +9,9 @@ var parse = require('./parse')
  */
 var addYears = function(dirtyDate, amount) {
   var date = parse(dirtyDate)
+  var time = getTimeSinceMidnight(date)
   date.setFullYear(date.getFullYear() + amount)
+  date.setHours(0, 0, 0, time)
   return date
 }