+++ src/set_day_of_year.js
@@ -1,4 +1,5 @@
 var parse = require('./parse')
+var getTimeSinceMidnight = require('./get_time_since_midnight')
 
 /**
  * Sets day of year to passed date.
@@ -8,8 +9,10 @@ var parse = require('./parse')
  */
 var setDayOfYear = function(dirtyDate, dayOfYear) {
   var date = parse(dirtyDate)
+  var time = getTimeSinceMidnight(date)
   date.setMonth(0)
   date.setDate(dayOfYear)
+  date.setHours(0, 0, 0, time)
   return date
 }