+++ src/set_year.js
@@ -1,4 +1,5 @@
 var parse = require('./parse')
+var getTimeSinceMidnight = require('./get_time_since_midnight')
 
 /**
  * Sets year to passed date.
@@ -8,7 +9,9 @@ var parse = require('./parse')
  */
 var setYear = function(dirtyDate, fullYear) {
   var date = parse(dirtyDate)
+  var time = getTimeSinceMidnight(date)
   date.setFullYear(fullYear)
+  date.setHours(0, 0, 0, time)
   return date
 }